Connected Settings to Backend
Added functionality to the settings in order for any changes made to the users account to be changed in the backend in Mongodb. Currently need to add some minor quality of life updates to the settings like aesthetics, clear the text field once an update is done, display the updated user information in the text field after a change was made, and make sure that the newpassword and confirmpassword are the same and authorized.